vmlinux.lds: Prevent LLD from eliminating built-in module versions
When the kernel is built with CONFIG_LD_DEAD_CODE_ELIMINATION=y and
linked with the LLD linker, the versions of built-in modules do not
appear in sysfs. One example is the "dummy" module - its defined version
is "1.0" which should show in /sys/module/dummy/version. However, with
LLD and DCE, this file does not exist.

Analysis using readelf shows that the __modver section, which module
versions are stored in, remains in the vmlinux binary, but has its type
changed to NOBITS as compared to PROGBITS:
LLD without DCE:
  [20] __modver          PROGBITS         ffffff80096e87e0  014b87e0
       0000000000000820  0000000000000000   A       0     0     8

LLD with DCE:
  [20] __modver          NOBITS           ffffff80094aa7e0  0137a7e0
       0000000000000820  0000000000000000   A       0     0     1

For some reason, LLD is ignoring the __used attribute present on the
declaration of the module_version_attribute structs. Instead of
depending on the linker to obey these annotations, explicitly tell the
linker to keep the entire __modver section instead in the linker script.
After all, the section is only 820 bytes.

The lack of a module version in sysfs also causes WireGuard apps to
fail to detect the presence of WireGuard in the kernel, though
WireGuard works fine if manually activated. This is because apps
check for WireGuard support using this sysfs node. This commit fixes
the issue by making sure that the /sys/module/wireguard/version node
exists.

Thanks to Nathan Chancellor <natechancellor@gmail.com> for help with
debugging.

Test: /sys/module/dummy/version exists, unlike before
Signed-off-by: kdrag0n <dragon@khronodragon.com>